Fixed #398: Changed datatype to be able to store > 255 second for backlight timer

+++ firmware/application/portapack.cpp
@@ -83,7 +83,7 @@ TransmitterModel transmitter_model;
 TemperatureLogger temperature_logger;
 
 bool antenna_bias { false };
-uint8_t bl_tick_counter { 0 };
+uint32_t bl_tick_counter { 0 };
 
 void set_antenna_bias(const bool v) {
 	antenna_bias = v;
